% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{fit_models}
\usage{
fit_models(data, k_range = 1:4, seed = NULL, weighted = FALSE)
}
\arguments{
\item{data}{data used to fit kmodes models. All variables must be numeric with no missing values.}

\item{k_range}{range of clustering solutions to fit. Defaults to 1:4}

\item{seed}{optional random seed. Defaults to \code{Null}}

\item{weighted}{Optional weighted version of distance algorithm.}
}
\value{
Tibble with results for each model


}
\description{
Fit x number of kmodes models specified with range of clustering solutions.
}
\examples{
fit_models(data, k_range = 4:6, seed = 1224, weighted = TRUE)



}
