% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elbow.R
\name{plot_elbow}
\alias{plot_elbow}
\title{plot_elbow}
\usage{
plot_elbow(models)
}
\arguments{
\item{models}{tibble created from \code{fit_models()}}
}
\value{
ggplot object
}
\description{
Creates an "elbow plot" of fit statistics for kmodes models.
}
\examples{
models <- fit_models(data)
          plot_elbow(models)


}
